Move top_k into _sorting.py
